lyrics_generator/pyproject.toml
[project]
name = "guitar-player-lyrics-generator"
version = "0.1.0"
description = "Lyrics transcription service using WhisperX (wav2vec2 forced alignment)"
requires-python = ">=3.10"
dependencies = [
    "whisperx>=3.1.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "boto3>=1.35.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "mangum>=0.19.0",
    "python-json-logger>=3.0.0",
    "openai>=1.0.0",
    "httpx>=0.27.0",
    "curl_cffi>=0.7.0",
    "beautifulsoup4>=4.12.0",
    "langdetect>=1.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.uv]
override-dependencies = [
    # Exclude torchcodec: transitive dep from pyannote-audio that has ABI
    # mismatch with torch 2.8. Not needed -- pyannote falls back to torchaudio.
    "torchcodec ; python_version < '0'",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lyrics_generator"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: tests that hit real external APIs (Genius, etc.)",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
pythonVersion = "3.11"
extraPaths = ["src"]
